requirements.txt
crewai
python-dotenv
requests
beautifulsoup4
langchain-community
langchain-openai
typer
rich
